Fix a bug where for some log messages that were supposed to be logged with a formatting template, the template itself was logged, followed by the formatting arguments, instead of the formatted message.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27226 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/core/algorithm/RestartAlgorithm.java
@@ -368,7 +368,7 @@ public class RestartAlgorithm implements Algorithm, StatisticsProvider, ReachedS
           if (!status.isSound()) {
             // if the analysis is not sound and we can proceed with
             // another algorithm, continue with the next algorithm
-            logger.log(
+            logger.logf(
                 Level.INFO,
                 "Analysis %d terminated, but result is unsound.",
                 stats.noOfAlgorithmsUsed);
@@ -376,7 +376,7 @@ public class RestartAlgorithm implements Algorithm, StatisticsProvider, ReachedS
           } else if (currentReached.hasWaitingState()) {
             // if there are still states in the waitlist, the result is unknown
             // continue with the next algorithm
-            logger.log(
+            logger.logf(
                 Level.INFO,
                 "Analysis %d terminated but did not finish: There are still states to be processed.",
                 stats.noOfAlgorithmsUsed);
